+++ statelit/utils/mro.py
@@ -18,9 +18,10 @@
 
 
 from enum import Enum
+from typing import Any
 from typing import Dict
-from typing import Iterable
 from typing import List
+from typing import Optional
 from typing import Tuple
 from typing import TypeVar
 from typing import Union
@@ -31,6 +32,7 @@ from pydantic.utils import lenient_issubclass
 
 
 T = TypeVar("T")
+Ty = TypeVar("Ty", bound=type)
 
 
 def extract_type_from_optional(cls: List[Union[type, Tuple[type, Tuple[type, ...]]]]) -> type:
@@ -100,8 +102,8 @@ def _c3_mro(cls, abcs=None):
     other_bases = list(cls.__bases__[boundary:])
     for base in abcs:
         if (
-            issubclass(cls, base)
-            and not any(issubclass(b, base) for b in cls.__bases__)
+            lenient_issubclass(cls, base)
+            and not any(lenient_issubclass(b, base) for b in cls.__bases__)
         ):
             # If *cls* is the class that introduces behaviour described by
             # an ABC *base*, insert said ABC to its MRO.
@@ -128,7 +130,7 @@ def _compose_mro(cls, types):
     # Remove entries which are already present in the __mro__ or unrelated.
 
     def is_related(typ):
-        return typ not in bases and hasattr(typ, '__mro__') and issubclass(cls, typ)
+        return typ not in bases and hasattr(typ, '__mro__') and lenient_issubclass(cls, typ)
 
     types = [n for n in types if is_related(n)]
     # Remove entries which are strict bases of other entries (they will end up
@@ -148,7 +150,7 @@ def _compose_mro(cls, types):
     for typ in types:
         found = []
         for sub in typ.__subclasses__():
-            if sub not in bases and issubclass(cls, sub):
+            if sub not in bases and lenient_issubclass(cls, sub):
                 found.append([s for s in sub.__mro__ if s in type_set])
         if not found:
             mro.append(typ)
@@ -162,7 +164,7 @@ def _compose_mro(cls, types):
     return _c3_mro(cls, abcs=mro)
 
 
-def _find_impl(cls: type, registry: Dict[type, T]) -> T:
+def _find_single_impl(cls: type, registry: Dict[type, T]) -> Optional[T]:
     """Returns the best matching implementation from *registry* for type *cls*.
 
     Where there is no registered implementation for a specific type, its method
@@ -181,40 +183,134 @@ def _find_impl(cls: type, registry: Dict[type, T]) -> T:
             if (
                     t in registry and t not in cls.__mro__
                     and match not in cls.__mro__
-                    and not issubclass(match, t)
+                    and not lenient_issubclass(match, t)
             ):
-                raise RuntimeError("Ambiguous dispatch: {} or {}".format(
-                    match, t))
+                raise RuntimeError(f"Ambiguous dispatch: {match} or {t}")
             break
         if t in registry:
             match = t
     return registry.get(match)
 
 
-def find_implementation(cls: type, registry: Dict[type, T]) -> T:
+def _find_all_impl(cls: Tuple[type, ...], registry: Dict[Tuple[type, ...], T]) -> Dict[Ty, T]:
+    mro = _compose_mro(cls, registry.keys())
+    matches = {}
+    for t in mro:
+        if t in registry:
+            matches[t] = registry[t]
+    return matches
+
+
+def _lenient_issubclass_respect_origin_plus_depth(cls, supercls, depth: int = 1) -> Tuple[int, bool]:
+    cls_origin = get_origin(cls)
+    supercls_origin = get_origin(supercls)
+
+    # treat `any` generically
+    if cls_origin and get_args(cls) in [(Any, ), (any, )]:
+        # treat any
+        cls = cls_origin
+        cls_origin = None
+    if supercls_origin and get_args(supercls) in [(Any, ), (any, )]:
+        # treat any
+        supercls = supercls_origin
+        supercls_origin = None
+
+    if cls_origin and supercls_origin:
+        if not lenient_issubclass(cls_origin, supercls_origin):
+            return depth, False
+        cls_args = get_args(cls)
+        supercls_args = get_args(supercls)
+        if not len(cls_args) == len(supercls_args):
+            return depth, False
+        _d = depth
+        for ca, sa in zip(cls_args, supercls_args):
+            _check_depth, check = _lenient_issubclass_respect_origin_plus_depth(ca, sa, depth=1)
+            _d += _check_depth
+            if not check:
+                return depth, False
+        else:
+            return _d, True
+    elif cls_origin is not None and supercls_origin is None:
+        # e.g. cls_origin == List[str] and supercls_origin == list
+        # in above example, cls_origin is superclass of supercls_origin
+        return _lenient_issubclass_respect_origin_plus_depth(cls_origin, supercls, depth=depth)
+    elif cls_origin is None and supercls_origin is not None:
+        return depth, False
+    elif supercls in (Any, any):
+        return depth, True
+    else:
+        # default to lenient_issubclass implementation.
+        return depth, lenient_issubclass(cls, supercls)
+
+
+def lenient_issubclass_respect_origin(cls, supercls):
+    return _lenient_issubclass_respect_origin_plus_depth(cls, supercls)[0]
+
+
+def find_implementation_traversing_origins_and_args(cls: type, registry: Dict[type, T]) -> Optional[T]:
+    # Motivting example:
+    #
+    # registry has:
+    # - list[dict]
+    # - list[dict[str, list]]
+    # - list[dict, [str, list[str]]
+    #
+    # if cls is list[dict, [str, list[int]], we want to return implementation
+    # for list[dict[str, list]].
+    #
+    # The solution is keep going deeper until we have exactly 1 impl.
+    #
+    # this approach is pretty inefficient, but it does work.
+    #
+    # care is taken to make sure that for simple lookups, inefficiency isn't a
+    # big deal, since complex typing should be very rare edge case. This is
+    # mainly here because, why not support complex types? :)
+    filtered_registry_keys = []
+    max_depth = -1
+    for k in registry:
+        depth, is_valid = _lenient_issubclass_respect_origin_plus_depth(cls, k)
+        if is_valid:
+            if depth > max_depth:
+                filtered_registry_keys = [k]
+                max_depth = depth
+            elif depth == max_depth:
+                filtered_registry_keys.append(k)
+            else:
+                pass
+    if len(filtered_registry_keys) == 1:
+        return filtered_registry_keys[0]
+    elif len(filtered_registry_keys) == 0:
+        return None
+    else:
+        raise RuntimeError(
+            f"There were {len(filtered_registry_keys)} implementations found that were valid:"
+            f" {filtered_registry_keys}."
+            "\n\n"
+            "In many cases, this error will be raised because MRO resolution is not supported"
+            " for registered implementations of complex types."
+            "\n\n"
+            "For example, if G is subtype of T, and List[G] and List[T] are both in the registry,"
+            " then List[G] will fail to resolve. (List[T] should work though.)"
+            "\n\n"
+            "If you want to support MRO resolution of complex types, help me out and contribute"
+            " to this library! :)"
+        )
+
+
+def find_implementation(cls: type, registry: Dict[type, T]) -> Optional[T]:
     cls = extract_type_from_optional(cls)
 
     origin = get_origin(cls)
     if origin is not None:
-        if issubclass(origin, tuple):
-            args = get_args(cls)
-            cls = args[0]
-            registry = {
-                k[1][0]: v
-                for k, v in registry.items()
-                if isinstance(k, tuple)
-                and len(k) == 2
-                and issubclass(k[0], tuple)
-                and isinstance(k[1], Iterable)
-            }
-        else:
-            cls = origin
+        # In the case that origin is not None, we avoid using the MRO.
+        k = find_implementation_traversing_origins_and_args(cls=cls, registry=registry)
+        return registry.get(k)
 
     # Give priority to Enum implementations
-    if issubclass(cls, Enum):
+    if lenient_issubclass(cls, Enum):
         _sub_registry = {k: v for k, v in registry.items() if lenient_issubclass(k, Enum)}
-        maybe_find_implementation = _find_impl(cls=cls, registry=_sub_registry)
+        maybe_find_implementation = _find_single_impl(cls=cls, registry=_sub_registry)
         if maybe_find_implementation:
             return maybe_find_implementation
 
-    return _find_impl(cls=cls, registry=registry)
+    return _find_single_impl(cls=cls, registry=registry)